move IconPick and JoinGameButton to home directory
