+++ src/home/HomeScreen.js
@@ -2,7 +2,7 @@ import { React, useState, useEffect } from 'react';
 import UserList from './UserList.js';
 import MainGameScreen from '../main/MainGameScreen.js';
 import PlayerStats from '../main/PlayerStats.js';
-import IconPick from '../IconPick';
+import IconPick from './IconPick';
 import {socket} from '../LoginScreen';
 import HomeButtons from './HomeButtons.js';
