+++ src/home/IconPick.js
@@ -3,7 +3,7 @@ import { Modal } from 'react-bootstrap';
 import 'emoji-mart/css/emoji-mart.css';
 import Button from 'react-bootstrap/Button';
 import { Picker,  Emoji } from 'emoji-mart';
-import {socket} from './LoginScreen';
+import {socket} from '../LoginScreen';
 
 function IconPick(){
     const [show, setShow] = useState(false);
